+++ CHANGELOG.md
@@ -35,6 +35,7 @@
   - Feat: typed parser.
   - Feat: built-in support for conflict resolver. [#7](https://github.com/DiscreteTom/retsac/issues/7)
   - Feat: in `AdvancedBuilder` you can rename literals.
+  - Feat: `children` in traverser is never `undefined`.
 - Optimize performance.
   - [#6](https://github.com/DiscreteTom/retsac/issues/6)
   - Lexer will skip unexpected actions by checking expected text.